Test/requirements.txt
behave==1.2.6
selenium==3.11.0
nose==1.3.7
PyYAML==3.12